+++ Source/astcenc_vecmathlib_sse_4.h
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: Apache-2.0
 // ----------------------------------------------------------------------------
-// Copyright 2019-2021 Arm Limited
+// Copyright 2019-2022 Arm Limited
 //
 // Licensed under the Apache License, Version 2.0 (the "License"); you may not
 // use this file except in compliance with the License. You may obtain a copy
@@ -412,7 +412,7 @@ ASTCENC_SIMD_INLINE vmask4 operator~(vmask4 a)
  */
 ASTCENC_SIMD_INLINE unsigned int mask(vmask4 a)
 {
-	return _mm_movemask_ps(a.m);
+	return static_cast<unsigned int>(_mm_movemask_ps(a.m));
 }
 
 // ============================================================================
@@ -801,7 +801,7 @@ ASTCENC_SIMD_INLINE vfloat4 round(vfloat4 a)
 	return vfloat4(_mm_round_ps(a.m, flags));
 #else
 	__m128 v = a.m;
-	__m128 neg_zero = _mm_castsi128_ps(_mm_set1_epi32(0x80000000));
+	__m128 neg_zero = _mm_castsi128_ps(_mm_set1_epi32(static_cast<int>(0x80000000)));
 	__m128 no_fraction = _mm_set1_ps(8388608.0f);
 	__m128 abs_mask = _mm_castsi128_ps(_mm_set1_epi32(0x7FFFFFFF));
 	__m128 sign = _mm_and_ps(v, neg_zero);
@@ -993,7 +993,7 @@ ASTCENC_SIMD_INLINE vfloat4 float16_to_float(vint4 a)
 ASTCENC_SIMD_INLINE float float16_to_float(uint16_t a)
 {
 #if ASTCENC_F16C >= 1
-	__m128i packed = _mm_set1_epi16(a);
+	__m128i packed = _mm_set1_epi16(static_cast<short>(a));
 	__m128 f32 = _mm_cvtph_ps(packed);
 	return _mm_cvtss_f32(f32);
 #else
@@ -1074,7 +1074,7 @@ ASTCENC_SIMD_INLINE vint4 vtable_8bt_32bi(vint4 t0, vint4 idx)
 {
 #if ASTCENC_SSE >= 30
 	// Set index byte MSB to 1 for unused bytes so shuffle returns zero
-	__m128i idxx = _mm_or_si128(idx.m, _mm_set1_epi32(0xFFFFFF00));
+	__m128i idxx = _mm_or_si128(idx.m, _mm_set1_epi32(static_cast<int>(0xFFFFFF00)));
 
 	__m128i result = _mm_shuffle_epi8(t0.m, idxx);
 	return vint4(result);
@@ -1096,7 +1096,7 @@ ASTCENC_SIMD_INLINE vint4 vtable_8bt_32bi(vint4 t0, vint4 t1, vint4 idx)
 {
 #if ASTCENC_SSE >= 30
 	// Set index byte MSB to 1 for unused bytes so shuffle returns zero
-	__m128i idxx = _mm_or_si128(idx.m, _mm_set1_epi32(0xFFFFFF00));
+	__m128i idxx = _mm_or_si128(idx.m, _mm_set1_epi32(static_cast<int>(0xFFFFFF00)));
 
 	__m128i result = _mm_shuffle_epi8(t0.m, idxx);
 	idxx = _mm_sub_epi8(idxx, _mm_set1_epi8(16));
@@ -1124,7 +1124,7 @@ ASTCENC_SIMD_INLINE vint4 vtable_8bt_32bi(vint4 t0, vint4 t1, vint4 t2, vint4 t3
 {
 #if ASTCENC_SSE >= 30
 	// Set index byte MSB to 1 for unused bytes so shuffle returns zero
-	__m128i idxx = _mm_or_si128(idx.m, _mm_set1_epi32(0xFFFFFF00));
+	__m128i idxx = _mm_or_si128(idx.m, _mm_set1_epi32(static_cast<int>(0xFFFFFF00)));
 
 	__m128i result = _mm_shuffle_epi8(t0.m, idxx);
 	idxx = _mm_sub_epi8(idxx, _mm_set1_epi8(16));
